Remove block code
