--- src/editor.scss
@@ -1,9 +1,0 @@
-/**
- * The following styles get applied inside the editor only.
- *
- * Replace them with your own styles or remove the file completely.
- */
-
-.wp-block-create-block-wp-ai-image-gen {
-	border: 1px dotted #f00;
-}
